requirements.txt
flask
flask-cors
pyaudio
wave
openai-whisper
vosk
ollama
